NGS-general: relocate 'split_fasta.py' into bcftbx.cli.
