+++ bin/split_fasta.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     split_fasta.py: extract individual chromosome sequences from fasta file
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.split_fasta import main
+if __name__ == "__main__":
+     main()
